Fix test for day 10 part 2

+++ 2020/src/day_10_part_2.rs
@@ -55,7 +55,7 @@ mod tests {
             ],
         ];
         let solutions = vec![8, 19208];
-        let outputs: Vec<u32> = inputs.iter().map(|v| super::solve(v.clone())).collect();
+        let outputs: Vec<u64> = inputs.iter().map(|v| super::solve(v.clone())).collect();
         assert_eq!(solutions, outputs);
     }
 }
